+++ src/Pathfinding.App.Console/Injection/Modules.cs
@@ -9,7 +9,6 @@ using Pathfinding.App.Console.ViewModels;
 using Pathfinding.App.Console.Views;
 using Pathfinding.Domain.Core.Enums;
 using Pathfinding.Domain.Interface.Factories;
-using Pathfinding.Infrastructure.Business;
 using Pathfinding.Infrastructure.Business.Algorithms;
 using Pathfinding.Infrastructure.Business.Algorithms.Heuristics;
 using Pathfinding.Infrastructure.Business.Algorithms.StepRules;
@@ -17,6 +16,7 @@ using Pathfinding.Infrastructure.Business.Commands;
 using Pathfinding.Infrastructure.Business.Layers;
 using Pathfinding.Infrastructure.Business.Serializers;
 using Pathfinding.Infrastructure.Business.Serializers.Decorators;
+using Pathfinding.Infrastructure.Business.Services;
 using Pathfinding.Infrastructure.Data.Pathfinding;
 using Pathfinding.Infrastructure.Data.Sqlite;
 using Pathfinding.Logging.Interface;
