+++ src/Pathfinding.Infrastructure.Business/Services/GraphInfoRequestService.cs
@@ -1,11 +1,11 @@
-using System.Linq;
-using Pathfinding.Domain.Interface;
 using Pathfinding.Domain.Interface.Extensions;
+using Pathfinding.Domain.Interface.Factories;
 using Pathfinding.Infrastructure.Data.InMemory;
 using Pathfinding.Service.Interface;
 using Pathfinding.Service.Interface.Models.Read;
+using Pathfinding.Shared.Extensions;
 
-namespace Pathfinding.Infrastructure.Business;
+namespace Pathfinding.Infrastructure.Business.Services;
 
 public sealed class GraphInfoRequestService(IUnitOfWorkFactory factory) : IGraphInfoRequestService
 {
