+++ src/Pathfinding.Infrastructure.Business/Services/MappingExtensions.cs
@@ -9,7 +9,7 @@ using Pathfinding.Service.Interface.Models.Undefined;
 using Pathfinding.Service.Interface.Requests.Create;
 using Pathfinding.Shared.Primitives;
 
-namespace Pathfinding.Infrastructure.Business;
+namespace Pathfinding.Infrastructure.Business.Services;
 
 internal static class MappingExtensions
 {
