+++ src/Pathfinding.Service.Interface/IStatisticsRequestService.cs
@@ -1,4 +1,3 @@
-using Pathfinding.Service.Interface.Models.Read;
 using Pathfinding.Service.Interface.Models.Undefined;
 using Pathfinding.Service.Interface.Requests.Create;
