+++ tests/Pathfinding.App.Console.Tests/ViewModelTests/RunTableViewModelTests.cs
@@ -8,6 +8,7 @@ using Pathfinding.Domain.Core.Enums;
 using Pathfinding.Infrastructure.Data.Pathfinding;
 using Pathfinding.Logging.Interface;
 using Pathfinding.Service.Interface;
+using Pathfinding.Service.Interface.Models.Read;
 using Pathfinding.Service.Interface.Models.Undefined;
 using Pathfinding.Shared.Extensions;
 using System.Reactive.Linq;
